feact(github actions): setup actions for forge

+++ src/mock/StandardTokenMock.sol
@@ -19,7 +19,7 @@ contract StandardTokenMock is ERC20 {
         _decimals = _decimal;
     }
 
-    function mint(address to, uint amount) external {
+    function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
